fix : config BeanUtil merged to common(#725)
