+++ hippo4j-common/src/main/java/cn/hippo4j/common/toolkit/BeanUtil.java
@@ -18,6 +18,8 @@
 package cn.hippo4j.common.toolkit;
 
 import cn.hippo4j.common.web.exception.IllegalException;
+import com.github.dozermapper.core.DozerBeanMapperBuilder;
+import com.github.dozermapper.core.Mapper;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.BeanUtils;
@@ -25,13 +27,47 @@ import org.springframework.beans.BeanUtils;
 import java.beans.IntrospectionException;
 import java.beans.PropertyDescriptor;
 import java.lang.reflect.Method;
-import java.util.Map;
+import java.util.*;
 
 /**
  * Bean util.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class BeanUtil {
+
+    protected static Mapper BEAN_MAPPER_BUILDER;
+
+    static {
+        BEAN_MAPPER_BUILDER = DozerBeanMapperBuilder.buildDefault();
+    }
+
+    public static <T, S> T convert(S source, Class<T> clazz) {
+        return Optional.ofNullable(source)
+                .map(each -> BEAN_MAPPER_BUILDER.map(each, clazz))
+                .orElse(null);
+    }
+
+    public static <T, S> List<T> convert(List<S> sources, Class<T> clazz) {
+        return Optional.ofNullable(sources)
+                .map(each -> {
+                    List<T> targetList = new ArrayList<T>(each.size());
+                    each.stream()
+                            .forEach(item -> targetList.add(BEAN_MAPPER_BUILDER.map(item, clazz)));
+                    return targetList;
+                })
+                .orElse(null);
+    }
+
+    public static <T, S> Set<T> convert(Set<S> sources, Class<T> clazz) {
+        return Optional.ofNullable(sources)
+                .map(each -> {
+                    Set<T> targetSize = new HashSet<T>(each.size());
+                    each.stream()
+                            .forEach(item -> targetSize.add(BEAN_MAPPER_BUILDER.map(item, clazz)));
+                    return targetSize;
+                })
+                .orElse(null);
+    }
 
     /**
      * copyProperties
