+++ hippo4j-config/pom.xml
@@ -50,17 +50,6 @@
         </dependency>
 
         <dependency>
-            <groupId>cn.hippo4j</groupId>
-            <artifactId>hippo4j-common</artifactId>
-            <version>${revision}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.github.dozermapper</groupId>
-            <artifactId>dozer-core</artifactId>
-        </dependency>
-
-        <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
         </dependency>
